% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_activity_model.R
\name{download_activity_model}
\alias{download_activity_model}
\title{Download Activity Model}
\usage{
download_activity_model(
  model = c("doherty-may20", "willetts-may20"),
  outdir = NULL,
  ...
)
}
\arguments{
\item{model}{Model to download for activity classification}

\item{outdir}{output directory for tarball}

\item{...}{additional arguments to pass to \code{\link{download.file}}}
}
\value{
The file path to the tarball
}
\description{
Download Activity Model
}
\examples{
download_activity_model(outdir = tempdir())
}
