% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{py_convert_cwa}
\alias{py_convert_cwa}
\alias{py_read_cwa}
\title{Read Axivity CWA file}
\usage{
py_convert_cwa(
  file,
  skipCalibration = TRUE,
  outputFolder = NULL,
  verbose = FALSE,
  timeZone = "Europe/London",
  timeShift = 0,
  startTime = NULL,
  endTime = NULL,
  epochPeriod = 30L,
  sampleRate = 100L,
  useFilter = TRUE,
  csvStartTime = NULL,
  csvSampleRate = NULL,
  calOffset = c(0, 0, 0),
  calSlope = c(1, 1, 1),
  calTemp = c(0, 0, 0),
  meanTemp = 20,
  stationaryStd = 13,
  rawDataParser = c("AccelerometerParser", "omconvert"),
  javaHeapSpace = NULL
)

py_read_cwa(file, ...)
}
\arguments{
\item{file}{the <.cwa/.cwa.gz> file to process
(e.g. sample.cwa.gz). If the file path contains
spaces,it must be enclosed in quote marks}

\item{skipCalibration}{skip calibration?}

\item{outputFolder}{folder for all of the output files}

\item{verbose}{enable verbose logging?}

\item{timeZone}{timezone in country/city format to
be used for daylight savings crossover check}

\item{timeShift}{time shift to be applied, e.g.
-15 will shift the device internal time by -15
minutes. Not to be confused with timezone offsets.}

\item{startTime}{removes data before this
time (local) in the final analysis}

\item{endTime}{removes data after this
time (local) in the final analysis}

\item{epochPeriod}{length in seconds of a single epoch}

\item{sampleRate}{resample data to n Hz}

\item{useFilter}{Filter ENMO trunc values?}

\item{csvStartTime}{start time for csv file
when time column is not available (e.g. 2020-01-01T00:01)}

\item{csvSampleRate}{sample rate for csv file
when time column is not available in Hz, or samples/second}

\item{calOffset}{accelerometer calibration offset}

\item{calSlope}{accelerometer calibration slope linking
offset to temperature}

\item{calTemp}{mean temperature in degrees Celsius of
stationary data for calibration}

\item{meanTemp}{mean calibration temperature in degrees}

\item{stationaryStd}{stationary mg threshold}

\item{rawDataParser}{type of parser to use.  If \code{omconvert}, this
must be installed in your environment path}

\item{javaHeapSpace}{amount of heap space (in MB) allocated to the
java subprocesses, useful for limiting RAM usage}

\item{...}{not used}
}
\value{
A list of files for the output
}
\description{
Read Axivity CWA file
}
\examples{
file =  system.file("extdata", "ax3_testfile.cwa.gz", package = "pycwa")
if (pycwa:::check_cwa_packages()) {
    res = pycwa::py_convert_cwa(file)
}
\donttest{
  if (pycwa:::check_cwa_packages()) {
    df = pycwa::py_read_cwa(file, startTime = "2019-02-26T10:55:15")
  }
}
}
