% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pycwa_required_modules.R
\docType{data}
\name{pycwa_required_modules}
\alias{pycwa_required_modules}
\title{Python Modules Required for \code{pycwa} Package}
\format{
A character vector of Python modules required to run the scripts
}
\usage{
pycwa_required_modules
}
\description{
Python Modules Required for \code{pycwa} Package
}
\keyword{datasets}
